backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.3.0
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.22
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
